+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Emitter::new()` and `Emitter::into_error()` to enable using the Emitter manually.
 - Implemented `Extend` for `Emitter` and `Error`.
 - Added `emit!` macro for adding errors to `Emitter`.
+- Added support for converting `darling::Error` to `manyhow::Error` (available via `darling` feature).
 
 ### Changed
 - **Breaking Change** replaced `Emitter::fail_if_dirty` with `Emitter::into_result`.